Replaced TreeNode specialization in favor of internal specialization
The specialization of TreeNode<DATA_T, -1, PTR_T> is removed as a whole. In place the specialization is done completely internally of the template class TreeNode by using the GetContainer<CHILDREN_COUNT> struct. The rest of the subroutines have been made aware of the vector specialization using constexpr if statements.